requirements.txt
azure-cli>=0.1.1b3
applicationinsights>=0.10.0
pip==9.0.1
prompt_toolkit==1.0.9
pygments==2.1.3
pylint==1.5.4
setuptools==30.4.0
pytest>=3.0.6
jmespath>=0.9.0